+++ src/main/java/com/example/repository/StreamRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by supercat on 24.4.17.
  */
-public interface StreamRepository extends JpaRepository<Stream, Long> {
+public interface StreamRepository extends JpaRepository<Stream, Integer> {
     
 }
